Moving round and hash back to the main file, and splitting ops in more files
