+++ ibis/backends/base/registry/__init__.py
@@ -1,6 +1,3 @@
-import itertools
-
-import ibis
 import ibis.common.exceptions as com
 import ibis.expr.datatypes as dt
 import ibis.expr.operations as ops
@@ -8,26 +5,16 @@ import ibis.expr.types as ir
 import ibis.util as util
 from ibis.backends.base_sqlalchemy import transforms
 
-from . import case, helpers, literal, window
-
-
-def binary_infix_op(infix_sym):
-    def formatter(translator, expr):
-        op = expr.op()
-
-        left, right = op.args
-
-        left_arg = translator.translate(left)
-        right_arg = translator.translate(right)
-        if helpers.needs_parens(left):
-            left_arg = helpers.parenthesize(left_arg)
-
-        if helpers.needs_parens(right):
-            right_arg = helpers.parenthesize(right_arg)
-
-        return '{} {} {}'.format(left_arg, infix_sym, right_arg)
-
-    return formatter
+from . import (
+    aggregate,
+    binary_infix,
+    case,
+    helpers,
+    literal,
+    string,
+    timestamp,
+    window,
+)
 
 
 def fixed_arity(func_name, arity):
@@ -40,72 +27,8 @@ def fixed_arity(func_name, arity):
     return formatter
 
 
-def identical_to(translator, expr):
-    op = expr.op()
-    if op.args[0].equals(op.args[1]):
-        return 'TRUE'
-
-    left_expr = op.left
-    right_expr = op.right
-    left = translator.translate(left_expr)
-    right = translator.translate(right_expr)
-
-    if helpers.needs_parens(left_expr):
-        left = helpers.parenthesize(left)
-    if helpers.needs_parens(right_expr):
-        right = helpers.parenthesize(right)
-    return '{} IS NOT DISTINCT FROM {}'.format(left, right)
-
-
-def xor(translator, expr):
-    op = expr.op()
-
-    left_arg = translator.translate(op.left)
-    right_arg = translator.translate(op.right)
-
-    if helpers.needs_parens(op.left):
-        left_arg = helpers.parenthesize(left_arg)
-
-    if helpers.needs_parens(op.right):
-        right_arg = helpers.parenthesize(right_arg)
-
-    return '({0} OR {1}) AND NOT ({0} AND {1})'.format(left_arg, right_arg)
-
-
 def unary(func_name):
     return fixed_arity(func_name, 1)
-
-
-def _reduction_format(translator, func_name, where, arg, *args):
-    if where is not None:
-        arg = where.ifelse(arg, ibis.NA)
-
-    return '{}({})'.format(
-        func_name,
-        ', '.join(map(translator.translate, itertools.chain([arg], args))),
-    )
-
-
-def reduction(func_name):
-    def formatter(translator, expr):
-        op = expr.op()
-        *args, where = op.args
-        return _reduction_format(translator, func_name, where, *args)
-
-    return formatter
-
-
-def variance_like(func_name):
-    func_names = {
-        'sample': '{}_samp'.format(func_name),
-        'pop': '{}_pop'.format(func_name),
-    }
-
-    def formatter(translator, expr):
-        arg, how, where = expr.op().args
-        return _reduction_format(translator, func_names[how], where, arg)
-
-    return formatter
 
 
 def not_null(translator, expr):
@@ -187,151 +110,6 @@ def log(translator, expr):
     return 'log({}, {})'.format(base_formatted, arg_formatted)
 
 
-def count_distinct(translator, expr):
-    arg, where = expr.op().args
-
-    if where is not None:
-        arg_formatted = translator.translate(where.ifelse(arg, None))
-    else:
-        arg_formatted = translator.translate(arg)
-    return 'count(DISTINCT {})'.format(arg_formatted)
-
-
-def substring(translator, expr):
-    op = expr.op()
-    arg, start, length = op.args
-    arg_formatted = translator.translate(arg)
-    start_formatted = translator.translate(start)
-
-    # Impala is 1-indexed
-    if length is None or isinstance(length.op(), ops.Literal):
-        lvalue = length.op().value if length is not None else None
-        if lvalue:
-            return 'substr({}, {} + 1, {})'.format(
-                arg_formatted, start_formatted, lvalue
-            )
-        else:
-            return 'substr({}, {} + 1)'.format(arg_formatted, start_formatted)
-    else:
-        length_formatted = translator.translate(length)
-        return 'substr({}, {} + 1, {})'.format(
-            arg_formatted, start_formatted, length_formatted
-        )
-
-
-def string_find(translator, expr):
-    op = expr.op()
-    arg, substr, start, _ = op.args
-    arg_formatted = translator.translate(arg)
-    substr_formatted = translator.translate(substr)
-
-    if start is not None and not isinstance(start.op(), ops.Literal):
-        start_fmt = translator.translate(start)
-        return 'locate({}, {}, {} + 1) - 1'.format(
-            substr_formatted, arg_formatted, start_fmt
-        )
-    elif start is not None and start.op().value:
-        sval = start.op().value
-        return 'locate({}, {}, {}) - 1'.format(
-            substr_formatted, arg_formatted, sval + 1
-        )
-    else:
-        return 'locate({}, {}) - 1'.format(substr_formatted, arg_formatted)
-
-
-def find_in_set(translator, expr):
-    op = expr.op()
-
-    arg, str_list = op.args
-    arg_formatted = translator.translate(arg)
-    str_formatted = ','.join([x._arg.value for x in str_list])
-    return "find_in_set({}, '{}') - 1".format(arg_formatted, str_formatted)
-
-
-def string_join(translator, expr):
-    op = expr.op()
-    arg, strings = op.args
-    return helpers.format_call(translator, 'concat_ws', arg, *strings)
-
-
-def string_like(translator, expr):
-    arg, pattern, _ = expr.op().args
-    return '{} LIKE {}'.format(
-        translator.translate(arg), translator.translate(pattern)
-    )
-
-
-def parse_url(translator, expr):
-    op = expr.op()
-
-    arg, extract, key = op.args
-    arg_formatted = translator.translate(arg)
-
-    if key is None:
-        return "parse_url({}, '{}')".format(arg_formatted, extract)
-    else:
-        key_fmt = translator.translate(key)
-        return "parse_url({}, '{}', {})".format(
-            arg_formatted, extract, key_fmt
-        )
-
-
-def extract_field(sql_attr):
-    def extract_field_formatter(translator, expr):
-        op = expr.op()
-        arg = translator.translate(op.args[0])
-
-        # This is pre-2.0 Impala-style, which did not used to support the
-        # SQL-99 format extract($FIELD from expr)
-        return "extract({}, '{}')".format(arg, sql_attr)
-
-    return extract_field_formatter
-
-
-def extract_epoch_seconds(t, expr):
-    (arg,) = expr.op().args
-    return 'unix_timestamp({})'.format(t.translate(arg))
-
-
-def truncate(translator, expr):
-    base_unit_names = {
-        'Y': 'Y',
-        'Q': 'Q',
-        'M': 'MONTH',
-        'W': 'W',
-        'D': 'J',
-        'h': 'HH',
-        'm': 'MI',
-    }
-    op = expr.op()
-    arg, unit = op.args
-
-    arg_formatted = translator.translate(arg)
-    try:
-        unit = base_unit_names[unit]
-    except KeyError:
-        raise com.UnsupportedOperationError(
-            '{!r} unit is not supported in timestamp truncate'.format(unit)
-        )
-
-    return "trunc({}, '{}')".format(arg_formatted, unit)
-
-
-def interval_from_integer(translator, expr):
-    # interval cannot be selected from impala
-    op = expr.op()
-    arg, unit = op.args
-    arg_formatted = translator.translate(arg)
-
-    return 'INTERVAL {} {}'.format(
-        arg_formatted, expr.type().resolution.upper()
-    )
-
-
-def null_literal(translator, expr):
-    return 'NULL'
-
-
 def value_list(translator, expr):
     op = expr.op()
     formatted = [translator.translate(x) for x in op.values]
@@ -395,48 +173,6 @@ def table_column(translator, expr):
     return quoted_name
 
 
-def timestamp_op(func):
-    def _formatter(translator, expr):
-        op = expr.op()
-        left, right = op.args
-        formatted_left = translator.translate(left)
-        formatted_right = translator.translate(right)
-
-        if isinstance(left, (ir.TimestampScalar, ir.DateValue)):
-            formatted_left = 'cast({} as timestamp)'.format(formatted_left)
-
-        if isinstance(right, (ir.TimestampScalar, ir.DateValue)):
-            formatted_right = 'cast({} as timestamp)'.format(formatted_right)
-
-        return '{}({}, {})'.format(func, formatted_left, formatted_right)
-
-    return _formatter
-
-
-def timestamp_diff(translator, expr):
-    op = expr.op()
-    left, right = op.args
-
-    return 'unix_timestamp({}) - unix_timestamp({})'.format(
-        translator.translate(left), translator.translate(right)
-    )
-
-
-def _from_unixtime(translator, expr):
-    arg = translator.translate(expr)
-    return 'from_unixtime({}, "yyyy-MM-dd HH:mm:ss")'.format(arg)
-
-
-def timestamp_from_unix(translator, expr):
-    op = expr.op()
-
-    val, unit = op.args
-    val = util.convert_unit(val, unit, 's').cast('int32')
-
-    arg = _from_unixtime(translator, val)
-    return 'CAST({} AS timestamp)'.format(arg)
-
-
 def exists_subquery(translator, expr):
     op = expr.op()
     ctx = translator.context
@@ -458,79 +194,54 @@ def exists_subquery(translator, expr):
     return '{} (\n{}\n)'.format(key, util.indent(subquery, ctx.indent))
 
 
-def nth_value(translator, expr):
+# XXX this is not added to operation_registry, but looks like impala is
+# using it in the tests, and it works, even if it's not imported anywhere
+def round(translator, expr):
     op = expr.op()
-    arg, rank = op.args
+    arg, digits = op.args
 
     arg_formatted = translator.translate(arg)
-    rank_formatted = translator.translate(rank - 1)
 
-    return 'first_value(lag({}, {}))'.format(arg_formatted, rank_formatted)
-
-
-def shift_like(name):
-    def formatter(translator, expr):
-        op = expr.op()
-        arg, offset, default = op.args
-
-        arg_formatted = translator.translate(arg)
-
-        if default is not None:
-            if offset is None:
-                offset_formatted = '1'
-            else:
-                offset_formatted = translator.translate(offset)
-
-            default_formatted = translator.translate(default)
-
-            return '{}({}, {}, {})'.format(
-                name, arg_formatted, offset_formatted, default_formatted
-            )
-        elif offset is not None:
-            offset_formatted = translator.translate(offset)
-            return '{}({}, {})'.format(name, arg_formatted, offset_formatted)
-        else:
-            return '{}({})'.format(name, arg_formatted)
-
-    return formatter
+    if digits is not None:
+        digits_formatted = translator.translate(digits)
+        return 'round({}, {})'.format(arg_formatted, digits_formatted)
+    return 'round({})'.format(arg_formatted)
 
 
-def ntile(translator, expr):
+# XXX this is not added to operation_registry, but looks like impala is
+# using it in the tests, and it works, even if it's not imported anywhere
+def hash(translator, expr):
     op = expr.op()
-    arg, buckets = map(translator.translate, op.args)
-    return 'ntile({})'.format(buckets)
+    arg, how = op.args
 
+    arg_formatted = translator.translate(arg)
 
-def day_of_week_index(t, expr):
-    (arg,) = expr.op().args
-    return 'pmod(dayofweek({}) - 2, 7)'.format(t.translate(arg))
-
-
-def day_of_week_name(t, expr):
-    (arg,) = expr.op().args
-    return 'dayname({})'.format(t.translate(arg))
+    if how == 'fnv':
+        return f'fnv_hash({arg_formatted})'
+    else:
+        raise NotImplementedError(how)
 
 
 binary_infix_ops = {
     # Binary operations
-    ops.Add: binary_infix_op('+'),
-    ops.Subtract: binary_infix_op('-'),
-    ops.Multiply: binary_infix_op('*'),
-    ops.Divide: binary_infix_op('/'),
+    ops.Add: binary_infix.binary_infix_op('+'),
+    ops.Subtract: binary_infix.binary_infix_op('-'),
+    ops.Multiply: binary_infix.binary_infix_op('*'),
+    ops.Divide: binary_infix.binary_infix_op('/'),
     ops.Power: fixed_arity('pow', 2),
-    ops.Modulus: binary_infix_op('%'),
+    ops.Modulus: binary_infix.binary_infix_op('%'),
     # Comparisons
-    ops.Equals: binary_infix_op('='),
-    ops.NotEquals: binary_infix_op('!='),
-    ops.GreaterEqual: binary_infix_op('>='),
-    ops.Greater: binary_infix_op('>'),
-    ops.LessEqual: binary_infix_op('<='),
-    ops.Less: binary_infix_op('<'),
-    ops.IdenticalTo: identical_to,
+    ops.Equals: binary_infix.binary_infix_op('='),
+    ops.NotEquals: binary_infix.binary_infix_op('!='),
+    ops.GreaterEqual: binary_infix.binary_infix_op('>='),
+    ops.Greater: binary_infix.binary_infix_op('>'),
+    ops.LessEqual: binary_infix.binary_infix_op('<='),
+    ops.Less: binary_infix.binary_infix_op('<'),
+    ops.IdenticalTo: binary_infix.identical_to,
     # Boolean comparisons
-    ops.And: binary_infix_op('AND'),
-    ops.Or: binary_infix_op('OR'),
-    ops.Xor: xor,
+    ops.And: binary_infix.binary_infix_op('AND'),
+    ops.Or: binary_infix.binary_infix_op('OR'),
+    ops.Xor: binary_infix.xor,
 }
 
 
@@ -563,17 +274,17 @@ operation_registry = {
     ops.DecimalPrecision: unary('precision'),
     ops.DecimalScale: unary('scale'),
     # Unary aggregates
-    ops.CMSMedian: reduction('appx_median'),
-    ops.HLLCardinality: reduction('ndv'),
-    ops.Mean: reduction('avg'),
-    ops.Sum: reduction('sum'),
-    ops.Max: reduction('max'),
-    ops.Min: reduction('min'),
-    ops.StandardDev: variance_like('stddev'),
-    ops.Variance: variance_like('var'),
-    ops.GroupConcat: reduction('group_concat'),
-    ops.Count: reduction('count'),
-    ops.CountDistinct: count_distinct,
+    ops.CMSMedian: aggregate.reduction('appx_median'),
+    ops.HLLCardinality: aggregate.reduction('ndv'),
+    ops.Mean: aggregate.reduction('avg'),
+    ops.Sum: aggregate.reduction('sum'),
+    ops.Max: aggregate.reduction('max'),
+    ops.Min: aggregate.reduction('min'),
+    ops.StandardDev: aggregate.variance_like('stddev'),
+    ops.Variance: aggregate.variance_like('var'),
+    ops.GroupConcat: aggregate.reduction('group_concat'),
+    ops.Count: aggregate.reduction('count'),
+    ops.CountDistinct: aggregate.count_distinct,
     # string operations
     ops.StringLength: unary('length'),
     ops.StringAscii: unary('ascii'),
@@ -584,40 +295,40 @@ operation_registry = {
     ops.LStrip: unary('ltrim'),
     ops.RStrip: unary('rtrim'),
     ops.Capitalize: unary('initcap'),
-    ops.Substring: substring,
+    ops.Substring: string.substring,
     ops.StrRight: fixed_arity('strright', 2),
     ops.Repeat: fixed_arity('repeat', 2),
-    ops.StringFind: string_find,
+    ops.StringFind: string.string_find,
     ops.Translate: fixed_arity('translate', 3),
-    ops.FindInSet: find_in_set,
+    ops.FindInSet: string.find_in_set,
     ops.LPad: fixed_arity('lpad', 3),
     ops.RPad: fixed_arity('rpad', 3),
-    ops.StringJoin: string_join,
-    ops.StringSQLLike: string_like,
+    ops.StringJoin: string.string_join,
+    ops.StringSQLLike: string.string_like,
     ops.RegexSearch: fixed_arity('regexp_like', 2),
     ops.RegexExtract: fixed_arity('regexp_extract', 3),
     ops.RegexReplace: fixed_arity('regexp_replace', 3),
-    ops.ParseURL: parse_url,
+    ops.ParseURL: string.parse_url,
     # Timestamp operations
     ops.Date: unary('to_date'),
     ops.TimestampNow: lambda *args: 'now()',
-    ops.ExtractYear: extract_field('year'),
-    ops.ExtractMonth: extract_field('month'),
-    ops.ExtractDay: extract_field('day'),
-    ops.ExtractQuarter: extract_field('quarter'),
-    ops.ExtractEpochSeconds: extract_epoch_seconds,
+    ops.ExtractYear: timestamp.extract_field('year'),
+    ops.ExtractMonth: timestamp.extract_field('month'),
+    ops.ExtractDay: timestamp.extract_field('day'),
+    ops.ExtractQuarter: timestamp.extract_field('quarter'),
+    ops.ExtractEpochSeconds: timestamp.extract_epoch_seconds,
     ops.ExtractWeekOfYear: fixed_arity('weekofyear', 1),
-    ops.ExtractHour: extract_field('hour'),
-    ops.ExtractMinute: extract_field('minute'),
-    ops.ExtractSecond: extract_field('second'),
-    ops.ExtractMillisecond: extract_field('millisecond'),
-    ops.TimestampTruncate: truncate,
-    ops.DateTruncate: truncate,
-    ops.IntervalFromInteger: interval_from_integer,
+    ops.ExtractHour: timestamp.extract_field('hour'),
+    ops.ExtractMinute: timestamp.extract_field('minute'),
+    ops.ExtractSecond: timestamp.extract_field('second'),
+    ops.ExtractMillisecond: timestamp.extract_field('millisecond'),
+    ops.TimestampTruncate: timestamp.truncate,
+    ops.DateTruncate: timestamp.truncate,
+    ops.IntervalFromInteger: timestamp.interval_from_integer,
     # Other operations
     ops.E: lambda *args: 'e()',
     ops.Literal: literal.literal,
-    ops.NullLiteral: null_literal,
+    ops.NullLiteral: literal.null_literal,
     ops.ValueList: value_list,
     ops.Cast: cast,
     ops.Coalesce: varargs('coalesce'),
@@ -625,19 +336,19 @@ operation_registry = {
     ops.Least: varargs('least'),
     ops.Where: fixed_arity('if', 3),
     ops.Between: between,
-    ops.Contains: binary_infix_op('IN'),
-    ops.NotContains: binary_infix_op('NOT IN'),
+    ops.Contains: binary_infix.binary_infix_op('IN'),
+    ops.NotContains: binary_infix.binary_infix_op('NOT IN'),
     ops.SimpleCase: case.simple_case,
     ops.SearchedCase: case.searched_case,
     ops.TableColumn: table_column,
     ops.TableArrayView: table_array_view,
-    ops.DateAdd: timestamp_op('date_add'),
-    ops.DateSub: timestamp_op('date_sub'),
-    ops.DateDiff: timestamp_op('datediff'),
-    ops.TimestampAdd: timestamp_op('date_add'),
-    ops.TimestampSub: timestamp_op('date_sub'),
-    ops.TimestampDiff: timestamp_diff,
-    ops.TimestampFromUNIX: timestamp_from_unix,
+    ops.DateAdd: timestamp.timestamp_op('date_add'),
+    ops.DateSub: timestamp.timestamp_op('date_sub'),
+    ops.DateDiff: timestamp.timestamp_op('datediff'),
+    ops.TimestampAdd: timestamp.timestamp_op('date_add'),
+    ops.TimestampSub: timestamp.timestamp_op('date_sub'),
+    ops.TimestampDiff: timestamp.timestamp_diff,
+    ops.TimestampFromUNIX: timestamp.timestamp_from_unix,
     transforms.ExistsSubquery: exists_subquery,
     transforms.NotExistsSubquery: exists_subquery,
     # RowNumber, and rank functions starts with 0 in Ibis-land
@@ -647,12 +358,12 @@ operation_registry = {
     ops.PercentRank: lambda *args: 'percent_rank()',
     ops.FirstValue: unary('first_value'),
     ops.LastValue: unary('last_value'),
-    ops.NthValue: nth_value,
-    ops.Lag: shift_like('lag'),
-    ops.Lead: shift_like('lead'),
+    ops.NthValue: window.nth_value,
+    ops.Lag: window.shift_like('lag'),
+    ops.Lead: window.shift_like('lead'),
     ops.WindowOp: window.window,
-    ops.NTile: ntile,
-    ops.DayOfWeekIndex: day_of_week_index,
-    ops.DayOfWeekName: day_of_week_name,
+    ops.NTile: window.ntile,
+    ops.DayOfWeekIndex: timestamp.day_of_week_index,
+    ops.DayOfWeekName: timestamp.day_of_week_name,
     **binary_infix_ops,
 }
