+++ ibis/backends/base/registry/binary_infix.py
@@ -1,0 +1,52 @@
+from . import helpers
+
+
+def binary_infix_op(infix_sym):
+    def formatter(translator, expr):
+        op = expr.op()
+
+        left, right = op.args
+
+        left_arg = translator.translate(left)
+        right_arg = translator.translate(right)
+        if helpers.needs_parens(left):
+            left_arg = helpers.parenthesize(left_arg)
+
+        if helpers.needs_parens(right):
+            right_arg = helpers.parenthesize(right_arg)
+
+        return '{} {} {}'.format(left_arg, infix_sym, right_arg)
+
+    return formatter
+
+
+def identical_to(translator, expr):
+    op = expr.op()
+    if op.args[0].equals(op.args[1]):
+        return 'TRUE'
+
+    left_expr = op.left
+    right_expr = op.right
+    left = translator.translate(left_expr)
+    right = translator.translate(right_expr)
+
+    if helpers.needs_parens(left_expr):
+        left = helpers.parenthesize(left)
+    if helpers.needs_parens(right_expr):
+        right = helpers.parenthesize(right)
+    return '{} IS NOT DISTINCT FROM {}'.format(left, right)
+
+
+def xor(translator, expr):
+    op = expr.op()
+
+    left_arg = translator.translate(op.left)
+    right_arg = translator.translate(op.right)
+
+    if helpers.needs_parens(op.left):
+        left_arg = helpers.parenthesize(left_arg)
+
+    if helpers.needs_parens(op.right):
+        right_arg = helpers.parenthesize(right_arg)
+
+    return '({0} OR {1}) AND NOT ({0} AND {1})'.format(left_arg, right_arg)
