--- ibis/backends/base/registry/unused.py
@@ -1,22 +1,0 @@
-def round(translator, expr):
-    op = expr.op()
-    arg, digits = op.args
-
-    arg_formatted = translator.translate(arg)
-
-    if digits is not None:
-        digits_formatted = translator.translate(digits)
-        return 'round({}, {})'.format(arg_formatted, digits_formatted)
-    return 'round({})'.format(arg_formatted)
-
-
-def hash(translator, expr):
-    op = expr.op()
-    arg, how = op.args
-
-    arg_formatted = translator.translate(arg)
-
-    if how == 'fnv':
-        return f'fnv_hash({arg_formatted})'
-    else:
-        raise NotImplementedError(how)
